+++ packages/react/src/components/controls/TrackToggle.tsx
@@ -14,7 +14,7 @@ export interface TrackToggleProps<T extends ToggleSource>
 }
 
 /**
- * With the TrackToggle component it is possible to mute and unmute your camera and microphone.
+ * With the `TrackToggle` component it is possible to mute and unmute your camera and microphone.
  * The component uses an html button element under the hood so you can treat it like a button.
  *
  * @example