+++ packages/react/src/components/layout/index.ts
@@ -1,6 +1,3 @@
 export * from './FocusLayout';
 export * from './GridLayout';
-export {
-  CarouselLayout as CarouselView,
-  CarouselLayoutProps as CarouselViewProps,
-} from './CarouselLayout';
+export { CarouselView, CarouselLayout, type CarouselLayoutProps } from './CarouselLayout';