+++ packages/react/src/components/participant/ParticipantTile.tsx
@@ -10,6 +10,7 @@ import {
   ParticipantContext,
   TrackRefContext,
   useEnsureParticipant,
+  useFeatureContext,
   useMaybeLayoutContext,
   useMaybeParticipantContext,
   useMaybeTrackRefContext,
@@ -22,7 +23,17 @@ import { AudioTrack } from './AudioTrack';
 import { useParticipantTile } from '../../hooks';
 import { useIsEncrypted } from '../../hooks/useIsEncrypted';
 
-/** @public */
+/**
+ * The `ParticipantContextIfNeeded` component only creates a `ParticipantContext`
+ * if there is no `ParticipantContext` already.
+ * @example
+ * ```tsx
+ * <ParticipantContextIfNeeded participant={trackReference.participant}>
+ *  ...
+ * </ParticipantContextIfNeeded>
+ * ```
+ * @public
+ */
 export function ParticipantContextIfNeeded(
   props: React.PropsWithChildren<{
     participant?: Participant;
@@ -69,7 +80,7 @@ export interface ParticipantTileProps extends React.HTMLAttributes<HTMLDivElemen
 }
 
 /**
- * The ParticipantTile component is the base utility wrapper for displaying a visual representation of a participant.
+ * The `ParticipantTile` component is the base utility wrapper for displaying a visual representation of a participant.
  * This component can be used as a child of the `TrackLoop` component or by passing a track reference as property.
  *
  * @example Using the `ParticipantTile` component with a track reference:
@@ -116,6 +127,8 @@ export function ParticipantTile({
   const isEncrypted = useIsEncrypted(p);
   const layoutContext = useMaybeLayoutContext();
 
+  const autoManageSubscription = useFeatureContext()?.autoSubscription;
+
   const handleSubscribe = React.useCallback(
     (subscribed: boolean) => {
       if (
@@ -144,7 +157,7 @@ export function ParticipantTile({
                 <VideoTrack
                   trackRef={trackReference}
                   onSubscriptionStatusChanged={handleSubscribe}
-                  manageSubscription={true}
+                  manageSubscription={autoManageSubscription}
                 />
               ) : (
                 isTrackReference(trackReference) && (