+++ packages/react/src/components/participant/TrackMutedIndicator.tsx
@@ -16,7 +16,8 @@ export interface TrackMutedIndicatorProps extends React.HTMLAttributes<HTMLDivEl
 }
 
 /**
- * The TrackMutedIndicator shows whether the participant's camera or microphone is muted or not.
+ * The `TrackMutedIndicator` shows whether the participant's camera or microphone is muted or not.
+ * By default, a muted/unmuted icon is displayed for a camera, microphone, and screen sharing track.
  *
  * @example
  * ```tsx